attempt4

+++ tasks/image_processor/grayscale.cpp
@@ -9,8 +9,9 @@ bool Grayscale::Apply(Bitmap *bitmap) {
     const double blue = 0.114;
     for (size_t i = 0; i < bitmap->GetHeight(); ++i) {
         for (size_t j = 0; j < bitmap->GetWidth(); ++j) {
-            unsigned char new_b = red * bitmap->GetPixels()(i, j).r + green * bitmap->GetPixels()(i, j).g +
-                                  blue * bitmap->GetPixels()(i, j).b;
+            unsigned char new_b =
+                static_cast<uint8_t>(red * bitmap->GetPixels()(i, j).r + green * bitmap->GetPixels()(i, j).g +
+                                     blue * bitmap->GetPixels()(i, j).b);
             unsigned char new_g = new_b;
             unsigned char new_r = new_b;
             bitmap->GetPixels()(i, j) = {new_b, new_g, new_r};
